requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Data processing
pandas>=2.1.0
tqdm>=4.66.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Database
psycopg2-binary>=2.9.9

# LangChain + OpenAI
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.10.0

# Pydantic
pydantic>=2.5.0

